moved common utils between brainmapper and brainreg readers to outside utils file
